feat(sidenav): high-contrast styles
